DSDEGP-1472 Make readGroupSchema JSON generation non-lazy. (#58)
I've been seeing the JSON schema test in ReadGroupWebServiceSpec
intermittently time out when run through our Dockerized SBT. The
timeout was arising because the ReadGroupWebService was failing to
respond to the schema query within 1 second. By making the schema ->
JSON transformation happen eagerly at startup, we reduce the amount
of work needed to be done when responding to the initial schema request,
(hopefully) avoiding the timeout.

It's hard to declare that the problem is definitively "fixed" by
this change, but after 100 runs of the test I haven't hit the
error (vs before, where I'd hit it a couple times in 100 runs).